+++ backsub/__main__.py
@@ -1,0 +1,4 @@
+from backsub.background_sub import main
+
+if __name__ == "__main__":
+    main()